emma-backend/requirements.txt
fastapi
uvicorn
pydantic
pyyaml
openai>=1.0.0
